+++ configs/typescript.cjs
@@ -1,0 +1,10 @@
+/** @type {import('eslint').Linter.Config} */
+module.exports = {
+  overrides: [{
+    files: "*.ts",
+    extends: [
+      "plugin:@typescript-eslint/recommended",
+      "plugin:@typescript-eslint/stylistic"
+    ]
+  }]
+};