+++ rulesets/typescript.cjs
@@ -1,4 +1,7 @@
 /** @type {import('eslint').Linter.Config} */
 module.exports = {
-  extends: "plugin:@typescript-eslint/recommended"
+  extends: [
+    "plugin:@typescript-eslint/recommended",
+    "plugin:@typescript-eslint/stylistic"
+  ]
 };