+++ rulesets/unicorn.cjs
@@ -2,8 +2,6 @@
 module.exports = {
   extends: "plugin:unicorn/recommended",
   rules: {
-    "unicorn/no-unsafe-regex": "error",
-    "unicorn/no-unused-properties": "error",
-    "unicorn/prefer-string-replace-all": "error"
+    "unicorn/no-unused-properties": "error"
   }
 };