+++ rulesets/vanilla.cjs
@@ -128,7 +128,6 @@ module.exports = {
     "no-redeclare": "error",
     "no-regex-spaces": "error",
     "no-return-assign": "error",
-    "no-return-await": "error",
     "no-script-url": "error",
     "no-sequences": "error",
     "no-shadow": "error",